+++ esp-hal/src/i2c/master/mod.rs
@@ -36,7 +36,13 @@ use fugit::HertzU32;
 use crate::{
     asynch::AtomicWaker,
     clock::Clocks,
-    gpio::{interconnect::PeripheralOutput, InputSignal, OutputSignal, Pull},
+    gpio::{
+        interconnect::{OutputConnection, PeripheralOutput},
+        InputSignal,
+        OutputSignal,
+        PinGuard,
+        Pull,
+    },
     interrupt::{InterruptConfigurable, InterruptHandler},
     pac::i2c0::{RegisterBlock, COMD},
     peripheral::{Peripheral, PeripheralRef},
@@ -439,6 +445,8 @@ pub struct I2c<'d, Dm: DriverMode> {
     phantom: PhantomData<Dm>,
     config: Config,
     guard: PeripheralGuard,
+    sda_pin: PinGuard,
+    scl_pin: PinGuard,
 }
 
 #[cfg(any(doc, feature = "unstable"))]
@@ -542,27 +550,35 @@ impl<'d, Dm: DriverMode> I2c<'d, Dm> {
     }
 
     /// Connect a pin to the I2C SDA signal.
-    pub fn with_sda(self, sda: impl Peripheral<P = impl PeripheralOutput> + 'd) -> Self {
+    ///
+    /// This will replace previous pin assignments for this signal.
+    pub fn with_sda(mut self, sda: impl Peripheral<P = impl PeripheralOutput> + 'd) -> Self {
         let info = self.driver().info;
         let input = info.sda_input;
         let output = info.sda_output;
-        self.with_pin(sda, input, output)
+        Self::connect_pin(sda, input, output, &mut self.sda_pin);
+
+        self
     }
 
     /// Connect a pin to the I2C SCL signal.
-    pub fn with_scl(self, scl: impl Peripheral<P = impl PeripheralOutput> + 'd) -> Self {
+    ///
+    /// This will replace previous pin assignments for this signal.
+    pub fn with_scl(mut self, scl: impl Peripheral<P = impl PeripheralOutput> + 'd) -> Self {
         let info = self.driver().info;
         let input = info.scl_input;
         let output = info.scl_output;
-        self.with_pin(scl, input, output)
+        Self::connect_pin(scl, input, output, &mut self.scl_pin);
+
+        self
     }
 
-    fn with_pin(
-        self,
+    fn connect_pin(
         pin: impl Peripheral<P = impl PeripheralOutput> + 'd,
         input: InputSignal,
         output: OutputSignal,
-    ) -> Self {
+        guard: &mut PinGuard,
+    ) {
         crate::into_mapped_ref!(pin);
         // avoid the pin going low during configuration
         pin.set_output_high(true);
@@ -572,9 +588,8 @@ impl<'d, Dm: DriverMode> I2c<'d, Dm> {
         pin.pull_direction(Pull::Up);
 
         input.connect_to(&mut pin);
-        output.connect_to(&mut pin);
 
-        self
+        *guard = OutputConnection::connect_with_guard(pin, output);
     }
 }
 
@@ -588,11 +603,16 @@ impl<'d> I2c<'d, Blocking> {
 
         let guard = PeripheralGuard::new(i2c.info().peripheral);
 
+        let sda_pin = PinGuard::new_unconnected(i2c.info().sda_output);
+        let scl_pin = PinGuard::new_unconnected(i2c.info().scl_output);
+
         let i2c = I2c {
             i2c,
             phantom: PhantomData,
             config,
             guard,
+            sda_pin,
+            scl_pin,
         };
 
         i2c.driver().setup(&i2c.config)?;
@@ -652,6 +672,8 @@ impl<'d> I2c<'d, Blocking> {
             phantom: PhantomData,
             config: self.config,
             guard: self.guard,
+            sda_pin: self.sda_pin,
+            scl_pin: self.scl_pin,
         }
     }
 
@@ -904,6 +926,8 @@ impl<'d> I2c<'d, Async> {
             phantom: PhantomData,
             config: self.config,
             guard: self.guard,
+            sda_pin: self.sda_pin,
+            scl_pin: self.scl_pin,
         }
     }