+++ esp-hal/src/uart.rs
@@ -125,9 +125,10 @@ use crate::{
     asynch::AtomicWaker,
     clock::Clocks,
     gpio::{
-        interconnect::{PeripheralInput, PeripheralOutput},
+        interconnect::{OutputConnection, PeripheralInput, PeripheralOutput},
         InputSignal,
         OutputSignal,
+        PinGuard,
         Pull,
     },
     interrupt::{InterruptConfigurable, InterruptHandler},
@@ -415,6 +416,9 @@ where
         let rx_guard = PeripheralGuard::new(self.uart.parts().0.peripheral);
         let tx_guard = PeripheralGuard::new(self.uart.parts().0.peripheral);
 
+        let rts_pin = PinGuard::new_unconnected(self.uart.info().rts_signal);
+        let tx_pin = PinGuard::new_unconnected(self.uart.info().tx_signal);
+
         let mut serial = Uart {
             rx: UartRx {
                 uart: unsafe { self.uart.clone_unchecked() },
@@ -425,6 +429,8 @@ where
                 uart: self.uart,
                 phantom: PhantomData,
                 guard: tx_guard,
+                rts_pin,
+                tx_pin,
             },
         };
         serial.init(config)?;
@@ -444,6 +450,8 @@ pub struct UartTx<'d, Dm> {
     uart: PeripheralRef<'d, AnyUart>,
     phantom: PhantomData<Dm>,
     guard: PeripheralGuard,
+    rts_pin: PinGuard,
+    tx_pin: PinGuard,
 }
 
 /// UART (Receive)
@@ -524,10 +532,10 @@ where
     Dm: DriverMode,
 {
     /// Configure RTS pin
-    pub fn with_rts(self, rts: impl Peripheral<P = impl PeripheralOutput> + 'd) -> Self {
+    pub fn with_rts(mut self, rts: impl Peripheral<P = impl PeripheralOutput> + 'd) -> Self {
         crate::into_mapped_ref!(rts);
         rts.set_to_push_pull_output();
-        self.uart.info().rts_signal.connect_to(rts);
+        self.rts_pin = OutputConnection::connect_with_guard(rts, self.uart.info().rts_signal);
 
         self
     }
@@ -536,12 +544,14 @@ where
     ///
     /// Sets the specified pin to push-pull output and connects it to the UART
     /// TX signal.
-    pub fn with_tx(self, tx: impl Peripheral<P = impl PeripheralOutput> + 'd) -> Self {
+    ///
+    /// Disconnects the previous pin that was assigned with `with_tx`.
+    pub fn with_tx(mut self, tx: impl Peripheral<P = impl PeripheralOutput> + 'd) -> Self {
         crate::into_mapped_ref!(tx);
         // Make sure we don't cause an unexpected low pulse on the pin.
         tx.set_output_high(true);
         tx.set_to_push_pull_output();
-        self.uart.info().tx_signal.connect_to(tx);
+        self.tx_pin = OutputConnection::connect_with_guard(tx, self.uart.info().tx_signal);
 
         self
     }
@@ -659,6 +669,8 @@ impl<'d> UartTx<'d, Blocking> {
             uart: self.uart,
             phantom: PhantomData,
             guard: self.guard,
+            rts_pin: self.rts_pin,
+            tx_pin: self.tx_pin,
         }
     }
 }
@@ -678,6 +690,8 @@ impl<'d> UartTx<'d, Async> {
             uart: self.uart,
             phantom: PhantomData,
             guard: self.guard,
+            rts_pin: self.rts_pin,
+            tx_pin: self.tx_pin,
         }
     }
 }
@@ -995,11 +1009,8 @@ impl<'d> Uart<'d, Blocking> {
     /// configure the driver side (i.e. the TX pin), or ensure that the line is
     /// initially high, to avoid receiving a non-data byte caused by an
     /// initial low signal level.
-    pub fn with_rx(self, rx: impl Peripheral<P = impl PeripheralInput> + 'd) -> Self {
-        crate::into_mapped_ref!(rx);
-        rx.init_input(Pull::Up);
-        self.rx.uart.info().rx_signal.connect_to(rx);
-
+    pub fn with_rx(mut self, rx: impl Peripheral<P = impl PeripheralInput> + 'd) -> Self {
+        self.rx = self.rx.with_rx(rx);
         self
     }
 
@@ -1007,13 +1018,8 @@ impl<'d> Uart<'d, Blocking> {
     ///
     /// Sets the specified pin to push-pull output and connects it to the UART
     /// TX signal.
-    pub fn with_tx(self, tx: impl Peripheral<P = impl PeripheralOutput> + 'd) -> Self {
-        crate::into_mapped_ref!(tx);
-        // Make sure we don't cause an unexpected low pulse on the pin.
-        tx.set_output_high(true);
-        tx.set_to_push_pull_output();
-        self.tx.uart.info().tx_signal.connect_to(tx);
-
+    pub fn with_tx(mut self, tx: impl Peripheral<P = impl PeripheralOutput> + 'd) -> Self {
+        self.tx = self.tx.with_tx(tx);
         self
     }
 }